data-processing/requirements.txt
python-frontmatter==1.0.0
PyYAML>=6.0
requests>=2.28.0
